+++ test/TorchSharpTest/TestTorchVisionOps.cs
@@ -6,9 +6,7 @@ using Xunit;
 
 namespace TorchSharp
 {
-#if NET472_OR_GREATER
     [Collection("Sequential")]
-#endif // NET472_OR_GREATER
     public class TestTorchVisionOps
     {
         [Fact]